243 refactored src file structure
