+++ src/digitalIdentity-frontend/src/app/pages/proofTemplate/create-proof-template/create-proof-template-stepper/auto-linked-attributes/auto-linked-attributes.component.ts
