+++ src/digitalIdentity-frontend/src/app/pages/proofTemplate/create-proof-template/create-proof-template-stepper/create-proof-template-step3b/create-proof-template-step3b.component.html
