+++ src/digitalIdentity-frontend/src/app/pages/proofTemplate/create-proof-template/create-proof-template-stepper/create-proof-template-step4/create-proof-template-step4.component.spec.ts
