+++ src/digitalIdentity-frontend/src/app/pages/proofTemplate/create-proof-template/create-proof-template-stepper/selected-cred-def-list/selected-cred-def-list.component.ts
