--- src/digitalIdentity-frontend/src/app/pages/proofTemplate/create-proofTemplate-page/create-proof-template/attribute.ts
@@ -1,6 +1,0 @@
-export interface Attribute {
-  attribID: number;
-  name: string;
-  value: string | number | Date;
-  type: 'String' | 'Number' | 'Email' | 'Date';
-}
