--- src/digitalIdentity-frontend/src/app/pages/proofTemplate/create-proofTemplate-page/create-proof-template/proof-template.ts
@@ -1,9 +1,0 @@
-export interface ProofTemplate {
-  name: string;
-  version: string;
-  credDefs: any[];
-  credDefStringAttributes: string;
-  credDefStringPredicates: string;
-  attributes: any[];
-  image: File | null;
-}
