--- src/digitalIdentity-frontend/src/app/pages/proofTemplate/create-proofTemplate-page/create-proofTemplate-page.component.html
@@ -1,174 +1,0 @@
-<mat-card class="create-proofTemplate-card">
-  <mat-card-header>
-    <mat-card-title class="create-proofTemplate-card-title">
-      Create a new proof template
-    </mat-card-title>
-  </mat-card-header>
-
-  <mat-divider [inset]="true"></mat-divider>
-
-  <mat-card-content class="create-proofTemplate-card-content">
-    <form
-      class="create-proofTemplate-form"
-      [formGroup]="proofTemplateFormGroup"
-    >
-      <mat-card-subtitle
-        >proof template information of
-        {{ proofTemplateFormGroup.value["name"] }}</mat-card-subtitle
-      >
-      <!-- <mat-form-field>
-        <mat-label>IconUrl:</mat-label>
-        <input
-          matInput
-          type="text"
-          placeholder="../../assets/images/DIdentity.png"
-          formControlName="iconUrl"
-        />
-      </mat-form-field> -->
-      <mat-form-field>
-        <mat-label>Name:</mat-label>
-        <input
-          matInput
-          type="text"
-          placeholder="Type in Name"
-          formControlName="name"
-        />
-      </mat-form-field>
-      <mat-form-field>
-        <mat-label>Version:</mat-label>
-        <input
-          matInput
-          type="text"
-          placeholder="Type in Version"
-          formControlName="version"
-        />
-      </mat-form-field>
-
-      <div>
-        <input
-          id="fileInput"
-          type="file"
-          hidden
-          (change)="selectFile($event)"
-          #inputFile
-        />
-        <button (click)="inputFile.click()" mat-raised-button disabled>
-          Select image
-        </button>
-        <mat-label *ngIf="error" style="color: red; text-align: center"
-          >&nbsp;&nbsp;{{ error }}</mat-label
-        >
-        <mat-label *ngIf="fileName"
-          >&nbsp;&nbsp;Image: {{ fileName }}</mat-label
-        >
-      </div>
-
-      <h3 class="create-proofTemplate-pick-credDefs">
-        Please select the required credential definitions:
-      </h3>
-
-      <div *ngIf="dataLoaded">
-        <app-filtered-table
-          (selectionChanged)="
-            this.selection = $event.dataSelection;
-            this.additionalData = $event.additionalData;
-            this.tableValid = $event.valid;
-            this.selectionChanged()
-          "
-          [data]="this.credDefData"
-          [internalColSelectNames]="this.selectableCols"
-          [displayedColSelectNames]="this.displayedColSelectNames"
-          [displayedColNames]="this.displayedColumnNames"
-          [internalColNames]="this.internalColumnNames"
-          [showExpandedDetails]="true"
-          [expandedDetails]="this.schemaDataAttributes"
-          [filterParams]="this.filterParams"
-        >
-        </app-filtered-table>
-      </div>
-
-      <mat-card-subtitle
-        >Requested attributes of proof template
-        {{ proofTemplateFormGroup.value["name"] }}</mat-card-subtitle
-      >
-      <div>
-        <button mat-raised-button (click)="addAttribute()">+</button>
-        &nbsp;&nbsp;&nbsp;
-        <mat-label>add requested attribute</mat-label>
-      </div>
-      <mat-form-field>
-        <mat-label>With Type:</mat-label>
-        <mat-select formControlName="nextType">
-          <mat-option *ngFor="let opt of types" [value]="opt">
-            {{ opt }}
-          </mat-option>
-        </mat-select>
-      </mat-form-field>
-      <div class="create-proofTemplate-form" formArrayName="attributes">
-        <!-- <img
-          src="{{ proofTemplateFormGroup.value['iconUrl'] }}"
-          alt="proofTemplateIcon"
-          class="logo1"
-        /> -->
-        <!-- <mat-grid-tile><img src="{{proofTemplateFormGroup.value['IconUrl']}}" alt="DIdentity" class="logo1"></mat-grid-tile> -->
-        <div *ngFor="let elem of attributes.controls; let i = index">
-          <mat-form-field>
-            <div [formGroupName]="i">
-              <mat-label
-                >{{ i }}:{{ proofTemplateTmp.attributes[i].type }}</mat-label
-              >
-              <input
-                matInput
-                formControlName="name"
-                id="{{ 'name' + i }}"
-                type="text"
-                placeholder="Enter Attribute Name"
-              />
-            </div>
-          </mat-form-field>
-          <mat-form-field [formGroupName]="i">
-            <mat-label>With Type:</mat-label>
-            <mat-select
-              formControlName="attributeType"
-              id="{{ 'attributeType' + i }}"
-              (selectionChange)="switchAttributeValue(i)"
-            >
-              <mat-option *ngFor="let opt of types" [value]="opt">
-                {{ opt }}
-              </mat-option>
-            </mat-select>
-          </mat-form-field>
-          <button (click)="deleteAttribute(i)">x</button>
-        </div>
-      </div>
-    </form>
-  </mat-card-content>
-  <button
-    mat-raised-button
-    [disabled]="
-      !tableValid ||
-      proofTemplateFormGroup.invalid ||
-      credentialDefinitionEmpty()
-    "
-    (click)="createProofTemplateButtonEvent()"
-  >
-    Save
-  </button>
-  <div *ngIf="!tableValid || proofTemplateFormGroup.invalid">
-    Form fields are not filled or filled with invalid arguments.
-  </div>
-  <div *ngIf="credentialDefinitionEmpty()">
-    No credential definition belongs to the proof template. You have to select
-    at least one credential definition.
-  </div>
-</mat-card>
-<br />
-<mat-card *ngIf="inDevelopment()">
-  <h2>Test output:</h2>
-  <!-- <p>iconUrl: {{ proofTemplate.iconUrl }}</p> -->
-  <p>name: {{ proofTemplate.name }}</p>
-  <p>version: {{ proofTemplate.version }}</p>
-  <div *ngFor="let elem of proofTemplate.attributes">
-    {{ elem.type }}:{{ elem.name }}
-  </div>
-</mat-card>
